source clean up 1st

+++ include/limit.h
@@ -4,13 +4,13 @@
 ************************************************************************* */
 
 /* Public Procedures */
-int mana_limit(struct char_data *ch);
-int hit_limit(struct char_data *ch);
-int move_limit(struct char_data *ch);
+long int mana_limit (struct char_data *ch);
+long int hit_limit (struct char_data *ch);
+long int move_limit (struct char_data *ch);
 
 struct title_type
-{
-	char *title_m;
-	char *title_f;
-	long int exp;
-};
+  {
+    char *title_m;
+    char *title_f;
+    long int exp;
+  };
